try to calculate the life total and failed however I am able to calculate the amount of damage and the number of hits which is all I really need

+++ paths/src/main/kotlin/bot/state/MapLocationState.kt
@@ -23,7 +23,7 @@ class MapLocationState(
 
     // remember for next interaction
     var previousLocation: FramePoint = FramePoint(),
-    var previousHeart: Int = -1,
+    var previousHeart: Double = -1.0,
     var previousGamePad: GamePad = GamePad.None,
 
     // the last known framestate
